- Cart Item Quantity Editable Done

+++ app/src/main/java/com/akg/akg_sales/dto/order/OrderRequest.java
@@ -12,15 +12,15 @@ public class OrderRequest {
     private Long customerId;
     private List<OrderLineRequest> lines = new ArrayList<>();
 
-    public void addLine(Long itemId,Double quantity){
+    public void addLine(Long itemId,Integer quantity){
         lines.add(new OrderLineRequest(itemId,quantity));
     }
 
     private class OrderLineRequest{
         private Long itemId;
-        private Double quantity;
+        private Integer quantity;
 
-        public OrderLineRequest(Long itemId, Double quantity) {
+        public OrderLineRequest(Long itemId, Integer quantity) {
             this.itemId = itemId;
             this.quantity = quantity;
         }
